AP_HAL_SITL: enable 32 servo outs

+++ libraries/AP_HAL_SITL/RCOutput.h
@@ -31,7 +31,7 @@ public:
     /*
       Serial LED emulation
      */
-    bool set_serial_led_num_LEDs(const uint16_t chan, uint8_t num_leds, output_mode mode = MODE_PWM_NONE, uint16_t clock_mask = 0) override;
+    bool set_serial_led_num_LEDs(const uint16_t chan, uint8_t num_leds, output_mode mode = MODE_PWM_NONE, uint32_t clock_mask = 0) override;
     void set_serial_led_rgb_data(const uint16_t chan, int8_t led, uint8_t red, uint8_t green, uint8_t blue) override;
     void serial_led_send(const uint16_t chan) override;
     
@@ -40,7 +40,7 @@ private:
     AP_ESC_Telem_SITL *esc_telem;
 
     uint16_t _freq_hz;
-    uint16_t _enable_mask;
+    uint32_t _enable_mask;
     bool _corked;
     uint16_t _pending[SITL_NUM_CHANNELS];
 };
